+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Process anyOf and allOf and well as oneOf https://github.com/OpenDataServices/compile-to-json-schema/issues/28
 
+### Removed
+
+- Removed Python 3.5 and 3.6 support, as they aren't supported any more
+
 ## [0.4.0] - 2021-07-30
 
 ### Added